Try to fix a 64-bit powerpc g++ 7 warning
Warning is:

allocate.hpp:335:19: warning: conversion to ???unsigned int??? from
???long unsigned int??? may alter its value [-Wconversion]

I'm not sure, but I think it's because the sizeof is a long unsigned
int, and the template parameter is an unsigned int. The sizeof isn't
even used, it's just there to get a value for expression SFINAE.

+++ include/boost/unordered/detail/allocate.hpp
@@ -325,7 +325,7 @@ namespace boost { namespace unordered { namespace detail {
 
 #if !defined(BOOST_NO_SFINAE_EXPR)
 
-    template <typename T, unsigned int> struct expr_test;
+    template <typename T, long unsigned int> struct expr_test;
     template <typename T> struct expr_test<T, sizeof(char)> : T {};
 
 #   define BOOST_UNORDERED_CHECK_EXPRESSION(count, result, expression)      \
